Convert crc32c ffi bindings to not use c2hs
Instead use direct bindings. The reason is this: the type of one of the
function arguments is C's type size_t. On some platforms this is an
alias for unsigned long, while on others it is unsigned long long. What
c2hs does is look through these alises and use either the Haskell type
CULong or CULLong. The problem then is that this causes a type error on
one platform or the other.

The solution is to use a Haskell type of CSize, which is guaranteed to
match the C type size_t. And if we do that, we don't need c2hs at all,
we can just use a simple direct binding.