Use Cabal cxx-sources and cxx-options for C++ code
Rather than cc-{sources,options}. This needs a later Cabal spec version,
so also update to 3.x.

This _should_ mean we pass the right c++ options to build the c++ code.
In particular this should mean we pass -std=c++11 which is needed on OSX
clang to get C++11 support, which the code needs to compile.