Add a couple asserts into the RawPage impl
These cause the tests to fail (when built with -fno-ignore-asserts).

The problem is that the tests currently assume that rawPageEntry will
work with pages with a single key/value pair in it. Arguably it should
but currently it will not. It only happens to work for because of the
little-endian representation, and would fail for values over 64k.

Furthermore the tests assume that they can pass over-long pages to
makeRawPage and then look up values and get them back in full. This
involves slicing outside of the bounds of the RawPage.