Change RawPage lookup to handle overflow case
Previously the rawPageEntry was designed to cover the case where
everything fit onto a single page (though it actually worked and was
tested for cases where it overflowed into immediately subsequent memory,
provided the value end offset remained within 16bits).

Now the renamed rawPageLookup can return either LookupEntryNormal or
LookupEntryOverflow. In the latter case it returns the value prefix and
it also returns the length of the trailing suffix that needs to be
fetched from other subsequent pages.

This means there's no pre-condition on rawPageLookup that it not be a
single-entry page.

Update the tests correspondingly.