Add a macrobenchmark of bloom filter lookup
With a setup resembling the LSM siutaion of many filters, in which we
look up the same key.

Also add an initial version of a bulk lookup for the same key in many
filters. This initial version just shares the hashing, but optimised
ones using memory prefetching are also possible.