Add the `Run` type to the fields of `BlobRef`
We have a 3 separate records for interfacing with the (real/simulated) file
system, and these 3 records should be shared across all runs and all table
handles within a session. These interfaces are now stored in the `Session` type.

A `BlobBref` now also has a reference to a `Run` in its fields, which provides
the blob file handle that will be read in blob retrieval. Blob references don't
have access to the file system interfaces, so `retrieveBlobs` now also requires
a Session argument.

Most of our code is parameterised over the monad `m`, and the type of file
handles `h`. To prevent the `h` parameter from escaping, we use existentials
inside the `Session` and public `BlobRef` type . An unfortunate consequence of
this is that we will now have to coerce/cast in `retrieveBlobs` to make the `h`
parameters match.

`Range` is moved to an internal module to prevent cyclic dependencies.

Make `TableHandle` hide an existential `h` like `Session` and `BlobRef` do,
using a new internal `AnySession` type as a reference to the `Session` that the
`TableHandle` was created in.