Extend RunAcc.new with a RangeFinderPrecision override
This will be needed for testing, where we want to test that the range
finder stuff works, but the normal policy would not start using range
finder bits until we get to very large sizes, which is awkward for
testing.

The previous PageAcc stuff that we're replacing has such an override for
its tests. So it's just mirroring that.