Fix the index-state for formatting jobs in GHA
Without a fixed index-state and with only one unique cache name before, GHA is
likely to rebuild dependencies from scratch each time we run the formatting
jobs. Fixing the index-state and using the value in the primary key of the GHA
caches, we won't rebuild dependencies superfluously any longer, and we have
control over when we want to build newer versions of our formatting tools, or
update their dependencies.