Extend the FormatPage reference impl with disk page sizes
So now instead of the model implementation producing arbitrary sized
pages, they're now padded to the underlying disk page size. The
prototype is aware of multiple disk page sizes, but we only use 4k to
compare against the real implementation.

This also now means that (in the model) encoding a page can fail,
because it can exceed the disk page size.

Several functions now have an extra DiskPageSize arg, and have Maybe
results. Update the other tests to match this API change.

We don't yet change any of the generators, or tests (other than for the
API changes).