`fs-sim` with simulated file I/O using user-supplied buffers
The `HasBufFS` primitives for performing file I/O using user-supplied buffers
are now merged into `HasFS`. This means we can get rid of the `HasBufFS`
parameter in a lot of places.

Both the `io-classes` and `io-sim` dependencies are now constrained to more recent
versions such that `IOSim` has a `PrimMonad` instance, which is now required for
using both the real and mocked `HasFS`.

`-fcheck-prim-bounds` is now enabled for both `fs-api` and `fs-sim`.