remove re-export of RunFsPaths
There is no need for a module re-export and we don't do it anywhere else.
Also, wanting to know the file paths to where a run stores its data
breaks the abstraction boundary and is only needed for tests or
benchmarks. As such, it's good to be more explicit about importing them.