Improve PageContent shrinkers and add shrink invariant tests
With the previous naive shrinkers for [(Key, Operation)] it was
infeasible to test the property that all shrinks satisfy an invariant
because there were simply too many possible shrinks (10's of 1000s).

It turns out that we can continue to use simple list shrinking for
[(Key, Operation)] provided that we use special case shrinking for Key
adn Value. These string types now use a shrinker that just tries three
things: take the first half, take all but the last byte, and replace
that last non-space char by a space. This "shrinking from the back"
tends to preserve sort order, while reducing size and simplifying
towards all trailing spaces. But by having only three alternatives, then
when used in shrinkers for larger structures then we don't end up with
so large a combinatorial explosion. This reduces the number of shrinks
for an arbitrary QC size 100 [(Key, Operation)] list by around a factor
of 250.

This then allows us to add prop_arbitraryAndShrinkPreserveInvariant
tests for invariants for PageContentFits and PageContentOrdered, without
those tests taking forever and too much memory.

Also adjust the "Shrinking satisfies invariant" property to sample from
the possible shrinks rather than testing them all. Note that this
optimisation alone was not enough to solve the problem with the original
naive shinker for [(Key, Operation)], because even generating the
shrinks -- to be able to sample from them -- took too much memory.