Make genPageContentSingle also make entries a bit bigger than disk page
The generator genPageContentSingleNearFull, used as one of the
components of genPageContentSingle, was only making results that were
slightly smaller than or equal to the disk page size. But for
single-entry pages it's legitimate to use multiple disk pages, and so we
should cover the case of slightly larger too, because it's a boundary
condition in the code that needs good coverage.