Add overflow page utilities to the FormatPage reference impl
And associated tests. There's a utility to return the number of overflow
pages expected, and if overflow pages are used, the length of the value
prefix and suffix within the first page and overflow pages.

These utilities will be helpful for comparing the real implementation
against the reference implementation (in particular extracting the value
prefix and suffix for the purpose of comparison).