Use the new generators in PageAcc tests
The PageAcc test needs to build pages, including making sure that the
boundary condition of the end of the page is covered well. So for this
test we use PageContentMaybeOverfull, since that produces key/op
sequences that include coverage of above and below the page size.

This revealed a bug, which is nice. The bug fix is in the next patch.