Fix a bug in PageAcc, and adjust size assumptions
The sizes for paOpMap and paBlobRefsMap were backwards. This caused
indexing out of bounds once one got to half the maximum number of keys.

Adjust the calculation of the maximum number of keys to simplify the
assumptions. Just assume that keys are non-empty, rather than assuming
they're all distinct.

To be safe (i.e. avoid indexing out of bounds) even in the case of all
empty keys, adjust the overflow test to also check the number of keys.
This means it fails to produce a result for such (extreme) test cases,
which is a difference versus the reference implementation, but this
difference can be handled in the tests explicitly without much
difficulty.