Make `RunFsPaths` aware of its parent directory
`RunFsPaths` is used to represent a run's set of file paths relative to the
session directory. However, it is not always the case that the session root is
the same as the `HasFS` root (i.e., `MountPoint`). For example, say we start a
session in "`MountPoint`/some/session/dir" (this is a textual representation of
an `FsPath`). In this case, our file-related code will fail, because
`RunFsPaths` thinks that the `MountPoint` is the session directory, not the
"some/session/dir".

The solution is to keep track of the session root `FsPath` somewhere. This
commit makes `RunFsPaths` aware of the directory that the run's files live in.
This means that `RunFsPaths` no longer represent paths relative to the session
root. Therefore, `RunFsPaths` also no longer prepends "active" to its paths. The
changes in this commit include:
* Add the run directory path to the `RunFsPath` type
* Interpret `RunFsPath`s as relative to this run directory field