Interim implementations of `snapshot` and `open`
This is a temporary implementation of the public API for taking and loading
snapshots, which we will use until we have properly implemented snapshotting.

The temporary snapshot format consists of a small file in the snapshots
directory that lists: (i) the runs that are used by a specific snapshot, and
(ii) its table configuration. We don't have to copy run files as part of the
snapshot because run files aren't (yet) deleted when runs are closed. The write
buffer is also persisted to disk as part of the snapshot, but the original table
remains unchanged.

`open` should be exception-safe, in the sense that it ensures no file resources
are leaked, even when some runs may have been opened already before an exception
is raised. We use a slightly generalised version of `Managed` from the @managed@
package to ensure resource cleanup in the presence of exceptions.

There is now a new `assertNoThunks` function that is similar to `assert`, but
with a custom error message in case an unexpected thunk is found.