Require `V*.` format for qualified imports of `vector` modules.
We use `hlint` to check whether the imports adhere to this format. For example:
```
import qualified Data.Vector as V
import qualified Data.Vector.Mutable as VM
import qualified Data.Vector.Unboxed as VU
import qualified Data.Vector.Unboxed.Mutable as VUM
```

Imports that did not adhere to this format in the `src/` and `test/` modules are
updated.