Proper resource cleanup in the lookup macro-benchmark
Asynchronous exceptions during the lookup macro-benchmark was not properly
cleaning up file resources. `withFS` now ensures that we do.

The benchmark directory for the lookup macro-benchmark is now also configurable.
If no filepath is provided, the system temporary directory will be used. If a
filepath is provided, then the user has to make sure that the path exists and
that it is a directory.