Use number of entries as the write buffer allocation method
We need a well-behaved measure of "run size" for merge scheduling. The number of
entries is such a well-behaved measure, whereas the byte/page size of a run is
not. For example, if entries vary much in terms of byte size, then the size in
pages of a run produced from the write buffer is hard to predict, and it is also
hard to predict the size of a run that is produced by merging runs. Moreover, we
need to know the number of entries (or keys) in each run for the Monkey-style
bit allocation to bloom filters.

Co-authored-by: Matthias Heinzel <matthias@well-typed.com>
Co-authored-by: Joris Dral <joris@well-typed.com>