Flush the write buffer when full
This is a first version that is functionally correct, but it has some
opportunities for optimisation.

We add a new `TempRegistry` abstraction to ensure proper cleanup for resources
that are bound to end up in some final state, after which they /should/ be
guaranteed to be released correctly. Example: the `TableContent` that is stored
in a `StrictMVar`.

The interim implementation of `snapshot` now reuses `flushWriteBuffer`, instead
of using its own local implementation of flushing.

`StateMachine` tests now perform some very naive shrinking, which was useful to
find some bugs in the current implementation of flushing.

Co-authored-by: Matthias Heinzel <matthias@well-typed.com>
Co-authored-by: Joris Dral <joris@well-typed.com>