Test for `FsResourceAlreadyExist` errors during session restore.
When a session is restored, we have to ensure we do not start the unique counter
at 0, because there might still be runs in the active directory that have names
larger than 0. When a run is created with the a name that already exists, we get
an `FsResourceAlreadyExist` error. This problem should go away once we have
proper snapshotting, but we have to provide a temporary fix until then.

The property currently fails, so we added an `expectFailure`. The next commit
fixes the bug.