Represent incoming runs on a `Level`, but it is always a fully merged run
Adding a `MergingRun` type to the representation of a `Level` means that the
code that performs merges (which is added later) can be a copy of the
`ScheduledMerges` prototype, with some minor changes to fit the real
implementation. However, since new merges will immediately be stepped to
completion until we have implemented scheduling, a `MergingRun` is only ever a
full run. We do, however, represent a single run differently from a run that was
the result of a merge. This is in line with the prototype as well, and might
make implementing the merge scheduling simpler.