QLS: enable tagging of the final state
The `tagStep` class function from `quickcheck-lockstep` allows for tagging each
individual step of the model, i.e., after each action that is run. However, in
some cases it is useful to tag a sequence of actions instead. For example, when
counting occurrences of a particular type, such as counting how many keys were
lookup over the course of running the sequence of actions. This commit adds an
alternative `runActionsBracket'` function that does just that, given a function
that tags the final state. Concrete tags are added in a latter commit.

Aside from the aforementioned change, this commit ensures that
`labelledExamples` no longer runs as part of the test suite. It can still be run
manually through a `repl` to inspect the `tagStep` function.