QLS: tune the generators
In particular, the generators are tuned to favour actions like `Updates` and
`Lookups` instead of actions like `New`, `Close` and Open`. This ensures that,
most of the time, we'll be running "interesting" actions with "interesting"
results. For example:

* If we `Close` too often, then it is likely that tables are closed when we
perform `Lookups` on them, which would just check that an error is returned
correctly instead of seeing lookup results.

* Generating `New` less often means that tables are more likely to evolve over
time through `Updates` before being `Close`d.