Add `hSetNoCache` to the `HasBlockIO` API
The `hSetNoCache` function can be used to set a data caching policy for
an open file handle. If the policy is set to no data caching on Linux,
FreeBSD or MacOS, reads and writes will try to bypass the OS page cache
as much as possible. This can be useful for the UTxO access pattern,
where there is little to no temporal locality for lookups. On Windows,
these functions are currently no-ops (although Windows does provide
similar functionality via the CreateFile flag FILE_FLAG_NO_BUFFERING).