Add optional testing to the WP8 benchmark.
Obviously we don't want to measure speed in this case, but it helps
improve confidence that it's calculating the right thing. This will be
especially useful for the pipelined version, which is non-trivial.

The approach is to insert output observaton into the benchmark, and
then optionally use it. The overhead of a no-op observation seems to
be less than the run-to-run variation.