Add WP8 benchmark pipelined mode
This is a bit awkward because to work with the real implementation it
requires the duplicate action, which isn't implemented yet.

I've tested the pipelined mode with the model implementation and that
works fine, but I don't want to include that in the benchmark code
because it pulls in too many other dependencies.

So to test the pipelined (and sequential) versions I've included a
dynamic check that the results are the same as a simple pure reference
function. This is enabled by --check. Hopefully the presence of the
observation/output doesn't have much of a performance impact when it is
disabled. Casual benchmark runsindicated the cost was less than the
run-to-run variation.