Restore a session's unique counter based on run numbers
Since a session's unique counter does not only determine the names of runs but
also serves as a source of unique identifiers for tables, the previous approach
was not sufficient to generate unique run names. This became apparent when using
the interim implementation of `duplicate`, which creates a host of new tables
with unique identifiers. Instead, we now read the names of the runs and pick a
strictly higher start value than those names for the unique counter when
restoring a session.