Add a custom `RWVar`
We need this primarily for the interim implementation of `duplicate`. We are
currently using the `RWVar` from `concurrent-extra`, but this hides internal
that we need to provide an interim implementation of `duplicate`.

What is also nice about this custom `RWVar` is that it is generic over the monad
`m` that it runs in. We technically do not need this yet, but it paves the way
for generalising our code to work for both `IO` and `IOSim`.

An `IOSimPOR` test is added, which is a rather trivial test, but I mainly added
it so I could get a little bit of practice with `IOSimPOR`.