Change the serialisation of `UTxOKey`
The hash part (`txId`) of a `UTxOKey` is a large, uniformly distributed hash,
and is therefore suitable to use in the compact index. However, the `txIx` part
of a `UTxOKey` is *not* uniformly distributed. This would lead to a high
probability of clashes in the compact index unless we incorporate the `txIx`
part in the prefix that is stored in the compact index. As such, the
serialisation format is now the following: `48` bits of `txId`, the full `16`
bits of `txIx`, the remaining `208` bits of `txId`.` The caveat is that ordering
is not preserved by the serialisation function, so results might be reported
slightly out of order by range lookups.