Move some Monkey functions to the monkey library
The moved functions now also use mainly `Integer`s instead of `Int` or `Word*`,
which means that some coercions between integral types have to happen. We use
`fromIntegral` to coerce to `Integer`, but we use `fromIntegralChecked` to
ensure that we don't silently overflow/underflow when we coerce back from an
`Integer`.

`unNumEntries` is now a separate function instead of a record field on
`NumEntries`, because this simplifies the `Show`/`Read` instances for this type.