Replace IORef by MutVar
As a result, quite a few types are now parameterised over a state parameter `s`
(e.g., `RealWorld` for `IO`). This change causes quite a bit of noise, but it's
necessary for generalising all our code away from just `IO`.

I have also considered whether it would be better to parameterise these types by
a monad `m` instead of the state type `s`. An `m` parameter is slightly more
general, since we can use `PrimState m` whenever `s` would be required. In the
spirit of the YAGNI principle, I decided not to do that (for now). We can make
that change later if turns out to be necessary.