Reset GHA caches for cabal stores
The size of a cabal store cache grows over time, since new dependencies are
built and added to the store, while old/unused dependencies are never removed
from the store. As such, we should "reset" our caches once in a while making
caches unreachable for restore/save actions, which we do in this commit by
changing the primary keys and restore keys.