Put `RWVar` and `TempRegistry` in a separate sublib
These components are general enough that they can be taken out of the core
library. This is also makes them more generally accessible for other components,
like the prototypes -- we are currently planning to prototype some concurrency
aspects of the main library.

Also, we add a TODO to the `TempRegistry` module about some renaming and
documentation updates.