Add new `RefCount` module
When we implement merge scheduling, we will have to be able to reference count
not only runs, but also ongoing merges. This commits adds a new `RefCount`
module that introduces a `RefCounter` type, which we can use to add reference
counting to a datatype.

This is based on the reference counter that the `Run` datatype already has, but
we factor it out so we can put a reference counter in other datatypes too. The
added benefit of factoring out the functionality instead of re-implementing it
for ongoing merges, is that we get uniform behaviour for all reference counted
objects. Another inspiration for the implementation is `Weak` pointer objects as
implemented by GHC.

The current design of `RefCounter m` is that it is put *inside* the type that is
reference counted. An alternative design I tried is to add a `RefCounted m a`
type that *wraps* a datatype a with reference counter behaviour. However, it has
a few disadvantages, such as adding an additional indirection, which causes
churn/noise and could have a performance impact. If we were to have a global
environment that tracks reference counts, then we might get around the
indirection, but then that environment has to be threaded through everywhere,
which seems tedious.