Add partial Internal.retrieveBlobs implementation
This is partial in that it does not support BlobRefs from the write
buffer, since those don't exist yet.

This implementation uses the HasBlockIO API to do all the reads for
all the blobs in one go. It also uses a single memory buffer for all the
I/O and then returns RawBytes that are slices of the single buffer.

The advantage is fewer separate allocations but a potantial disadvantage
is that all the blobs looked up will have the same lifetime. This design
choice could be changed easily if we decide independent lifetimes is
better.