Change BlobRef representation to not be specific to Run
The goal is to have a single representation for BlobRefs whether they're
from runs or from the write buffer.

The intention is that this representation will also be sufficient to
cover BlobRefs for blobs from the write buffer, using an open file
handle and a suitable reference count.

So rather than containing a run it contains a file handle and a
reference count.

- blobRefRun   :: !run
+ blobRefFile  :: !h
+ blobRefCount :: {-# UNPACK #-} !(RefCounter m)

When making a BlobRef for a run, we just take the blob file handle and
the run's reference count.