Model exceptions in `retrieveBlobs` properly.
The modelled `retrieveBlobs` would throw a "table is closed" error if a table
corresponding to a blob references is closed, but now it will return a "blob
reference invalid" error.

A slightly unrelated change is that we generalise the model `BlobRef` types, so
that `Cursor`s can also return `BlobRefs`. Cursors will be added to the model in
subsequent commits.