Clarify distinction between weak and strong BlobRefs
Introduce an internal WeakBlobRef type. This is used as the internal
representation for the public API BlobRef type.

Otherwise internally, WeakBlobRef is the weak version and BlobRef is the
strong one where you have (or are assumed to have) a reference on the
file/run whatever.

And then provide utilities for dereferencing a weak ref to get a strong
ref. This means we get all the weak refcount stuff in one place (in the
BlobRef module) and its use in Database.LSMTree.Internal becomes quite
straightforward and doesn't need access to internal representations.