Stop using WriteBuffer type for test data so much
Many tests rely on the WriteBuffer type as a convenient represntation
for generated input test data, since it's a bundle of key/operation
pairs.

The representation and API of the WriteBuffer is about to change,
however which will break all of these uses. So we make two changes:
1. change TypedWriteBuffer to contain a Map of an appropriate type
   rather than a WriteBuffer
2. introduce a utility function mkRunFromSerialisedKOps which takes a
   Map of an appropriate type rather than a WriteBuffer

This will insulate these tests from the upcoming changes in the
WriteBuffer. The mkRunFromSerialisedKOps implementation will need to
change, but it's interface can remain stable.