Add new WriteBufferBlobs abstraction for on disk blob storage
This is intended to support converting the WriteBuffer to store just
blof references, while keeping the blobs themselves on disk. Blobs are
suppsoed to be written out to disk upon insertion, rather than when the
write buffer is flushed to a run.

A WriteBufferBlobs is intended to be shared between many table handles,
to support cheap duplication of tale handles. This constraint informs
several aspects of its design.

It is also intended to support a single shared representation of
BlobRefs for blobs from runs or from the write buffer. In particular
this includes a reference count to manage the lifetime of the write
buffer blob file.