Wire up BlobRefs from write buffer lookups
Previously WB.lookup would return type WeakBlobRef m h (which is the
convenient type), but would actually throw an error. Now it returns
BlobSpan but works. The caller then needs to convert from BlobSpan to
WeakBlobRef (and then wrap that into the public API's BlobRef type).

Also adjust the WBB.mkBlobRef to be a monadic action returning a pure
function. This makes it more convenient to use in the lookup code to
convert results using BlobSpan into results using WeakBlobRef. This
works but it's not terribly pretty.

Some reorganisation here would be good, to reduce and combine
conversions to reduce allocations.