Enable Blobs in many tests, including StateMachine tests
Only one fix in the code is needed, for the NoThunks test.

The state machine test found two non-trivial differences between the
spec and the model in relation to blobs:
1. blob retrieval after snapshot fails
2. blob retrieval after failed-snapshot (due to duplicate name) fails

The reason for both of these is that the implementation currently works
by modifying flushing the write buffer to disk, and modifing the table
handle to use the updated set of runs and now-empty write buffer. This
invalidates all blob references from the write buffer itself.

For the moment I have "fixed" this difference by changing the model to
behave like the implementation. For the TODO task to fix this, see
https://github.com/IntersectMBO/lsm-tree/issues/392