Properly reference count the WriteBufferBlobs in cursors
Cursors, being like read-only tables, need to do the same reference
counting for the WriteBufferBlobs as tables do. In particular they need
to take a reference on the blob store when the cursor is created (in
case the original table is modified or closed), and release the
reference when the cursor itself is closed.