Refactor: push write buffer lookups into lookupsIO
Previously lookupsIO dealt only with lookups in runs, and combining
results from the write buffer was done in D.L.Internal. Originally that
was ok because combining the write buffer results was simple, but it has
grown more complex with the blobs feature being properly enabled.

So now we pass it down into lookupsIO, specifically into
intraPageLookups where all the results are combined. It works nicely
here, since per-key results are accumulated into a mutable vector, so
the only change is to initialise that array with the write buffer
lookups. All the lookup results from runs are then accumulated on top.
This should also mildly reduce allocations.

In addition, instead of D.L.Internal.lookups getting passed in a
function for converting into the final result, just return the
intermediate representation and have the final conversion done in the
API wrappers. This approach is more friendly for inlining and
specialisation (so the deserialisation can be specialised at the call
site).

The benchmarks are updated but use an empty write buffer. Plausibly we
might want some micro benchmarks for lookupIO et al that do make use of
the write buffer too.