Make `IndexCompact.search` return valid page spans in some edge cases
In case `search` was given a key that was smaller than all keys on all pages in
the index, then `search` would return `PageSpan 0 0` by default. However, if
page 0 contains a large value that overflows into subsequent pages, then
`PageSpan 0 0` is not valid because we have to include all the overflow pages as
well. Instead, `search` now consults the `LTP` array to ensure that we return a
valid page span. In addition, the `search` code has been cleaned up a little,
and it's documentation has been improved.

A `TODO` in `RunReader.new` has been addressed, which depended on `search`
returning valid page spans.