Fix bug where `RWVar` was not actually biased towards writers
A `retry` in the wrong place was causing writers to wait for all readers to
finish, even as new readers were added. The `RWVar` should now be more properly
biased towards writers, though there is a new TODO in the code that describes
how a `RWVar` could be made to be even more biased.