Store `HasFS` and `HasBlockIO` in the `Merge` type
This leads to a minor simplification of the code, because we avoid explicitly
passing `HasFS` and `HasBlockIO` around in some places. As a side-effect, we
have specialised some of `Merge`'s record fields to use file handles of type
`Handle h`, instead of some arbitrary `fhandle`. This was implicit before, since
`Handle h` is the only type we can use in combination with `HasFS` and
`HasBlockIO`, but now it is hardcoded into the type.