Separate merge *completion* from merge *stepping*
When a merge is completed, its output is used to create a new `Run`, and in the
process some system resources such as file handles will be created/deleted. To
prevent leaks of these system resources, we have to be careful with where we
allow asynchronous exceptions to happen, and so we will want to `mask` (part of)
the completion phase.

Previously, the completion phase was incorporated into the function `steps`,
which was also in charge of doing the main merging work, i.e., consuming entries
from input runs and writing them to an output run. However, we do not want to
`mask` the whole of the `steps` function.

By factoring out the completion phase into a separate `complete` function, we
can do more fine-grained `mask`ing.