Add new `RunData` datatype and utilities.
This datatype is intended to replace the `TypedWriteBuffer` datatype. A
`RunData` represents a collection of key\/value pairs that is suitable for
creating a `Run`. Along with generators, and a labelling function, the new
module provides `bracket`-style functions that can be used to temporarily create
a run from `RunData`.