Enable blobs in test prop_roundtripFromWriteBufferLookupIO
This test creates a write buffer and bunch of runs and then does
lookups in the combination. So this nicely tests the new lookupsIO
approach that includes the write buffer. And now it also covers
checking the blobs we get out are the ones expected. So this should
cover getting blobs both from the write buffer and from runs.