Adjust RunReaders lock-step test to allow popping more than available
Previously the test case generation had an (implicit) invariant that it
would not try to pop more elements than were available. This invariant
was not preserved by shrinking however. This resulted in assertion
failures when shrinking other failing test cases.

Although the lock-step infrastructure can preserve invariants during
shrinking, we can do something simpler here: just allow popping more
elements than available and tests that this gives the expected (failing)
result. This required a change to the model to drain all elements when
too many were asked for, and remove an assertion from the interpreter
for the real implementation. We also extend the generator to explicitly
cover the case of popping too many elements.