`NoThunks` fix for `PageAcc`
`PageAcc` can contain uninitialised memory in the vectors for keys and values,
and its `NoThunks` instance will throw errors when that memory is checked. In
this commit, we ensure that the vectors are filled in with dummy keys and
values. Moreover, when a `PageAcc` is reset, we ensure that existing keys and
values are cleared and replaced by dummy keys and values.