Remove deferred open feature from WriteBufferBlobs
It was not safe for concurrent use (as is required) and it is now
unclear if it would be any great saving. The main original intention was
to make table handle duplication cheap, but that remains cheap since the
WriteBufferBlobs can be shared without doing anything other than
incrementing a reference count.