Refactor RunBuilder.addKeyOp to include copying blobs
Rather than all callers copying the blobs first and passing in BlobSpans
for the target, we just pass in the BlobRef of the source, and then
addKeyOp copies the blob.

We could still do better in terms of number of allocations. The scheme
with Entry involves quite a bit of reallocation because of the need to
convert from BlobSpan to BlobRef and back.