Run lockstep tests with `IOSim`
This has some implications for `retrieveBlobs`: we remove the `Typeable m`
constraint, which is not satisfiable for `IOSim s` due to the state type `s`,
which is polymorphic. It turns out we do not need that constraint anyway,
because it is only the existential type variables for file handles have to be
checked for type equality.