Run lockstep tests with simulated `HasFS` and `HasBlockIO`
With the nice addition that we now assert that there are no open file handles at
the end of each iteration of the tests!

We also tweak the state machine tests. We now follow a more normal convention,
and allow running individual tests more easily in GHCi via `quickCheck $
theProperty`.

Co-authored-by: Duncan Coutts <duncan@well-typed.com>
Co-authored-by: Joris Dral <joris@welltyped.com>