Fix adding/removing references on table content
In some places, we were adding/removing references but not taking into account
rollbacks/delays that should be performed when exceptions occur. In particular,
correct acquisition and release of write buffer blob files were not guarded by
`allocateTemp` and `freeTemp`.

In addition, to make our reference accounting slightly less error prone, we make
a few related changes:

* We add new  `addReferenceTableContent` and `removeReferenceTableContent`
functions, which replace most occurrences of manually increasing/decreasing
reference counts for the components of a  `TableContent`. For example, in
`duplicate` and `close` we now use the new functions.

* `newWith` now uses a `TempRegistry`, which fixes a double-free TODO.

* `newWith` now requires a full `TableContent` to be passed in.