WIP: Add `removeReferenceN`
In some cases, we have to remove more than one reference at a time. For example,
when a merge is completed, then the reference count of each input run should be
reduced by value of the reference count of the output run. You can view this as
the references being "transferred" from the input runs to the output run.