Simulated `tryLockFile` now keeps open a read handle until it is unlocked
Before this change, a successfull `tryLockFile` would not contribute to the
number of open file handles. Instead, it would open a read-write handle to
increment/decrement a counter in the locked file, after which the handle is
released again before returning a `LockFileHandle`. To check whether a file has
been fully unlocked, we would have to read the contents of the locked file,
which is more complicated than it has to be. Instead, we now ensure that
`tryLockFile` acquires a read handle after updating the counter in the file, and
this handle is only released after calling `hUnlock` on the `LockFileHandle`. As
such, a file lock will contribute to the number of open file handles in a
`MockFS`. Our tests can then simply read this number, and make assertions about
the number.