QLS tests: label the number of tables and number of actions per table
In the QLS tests.

Initial results
      Number of tables (100 in total):
      33% NumTables "2 <= n < 4"
      23% NumTables "4 <= n < 8"
      17% NumTables "1 <= n < 2"
      16% NumTables "8 <= n < 16"
       9% NumTables "n == 0"
       2% NumTables "16 <= n < 32"

      Number of actions per table (410 in total):
      27.8% NumTableActions "n == 0"
      22.0% NumTableActions "2 <= n < 4"
      19.3% NumTableActions "4 <= n < 8"
      16.1% NumTableActions "1 <= n < 2"
      12.4% NumTableActions "8 <= n < 16"
       2.4% NumTableActions "16 <= n < 32"

As we can see, 9% of runs have no tables at all! This seems a bit
surprising, and not terribly useful. Although such runs will be quick,
in CI we typically only run 100 tests.

And also in many runs there a lot of tables (nearly 20% have 8 or more).
This is probably too many to get interesting interleavings of operations
between related (duplicated) tables. So we may want to reduce the number
of tables, and make more related tables so we can get better coverage of
interleaved operations on related tables.

And as we can also see, over a quarter of tables have no actions at all!
And generaly the number of actions per table is quite low. This would
probably be better with fewer tables.